backend/requirements.txt
# django requirements
django
whitenoise
azure-monitor-opentelemetry
psycopg2
ipython
django-debug-toolbar
azure-storage-blob
